Support project option for prettier
Massive refactor, sorry I didn't split the commits
