+++ packages/jest-config/src/defaults.js
@@ -61,6 +61,7 @@ export default ({
   notify: false,
   notifyMode: 'always',
   preset: null,
+  prettier: 'prettier',
   projects: null,
   resetMocks: false,
   resetModules: false,
