+++ packages/jest-config/src/valid_config.js
@@ -65,6 +65,7 @@ export default ({
   notifyMode: 'always',
   onlyChanged: false,
   preset: 'react-native',
+  prettier: '<rootDir>/node_modules/prettier',
   projects: ['project-a', 'project-b/'],
   reporters: [
     'default',
