+++ packages/jest-jasmine2/src/setup_jest_globals.js
@@ -100,7 +100,11 @@ export default ({
     });
   patchJasmine();
   const {expand, updateSnapshot} = globalConfig;
-  const snapshotState = new SnapshotState(testPath, {expand, updateSnapshot});
+  const snapshotState = new SnapshotState(testPath, {
+    expand,
+    getPrettier: () => localRequire(config.prettier),
+    updateSnapshot,
+  });
   setState({snapshotState, testPath});
   // Return it back to the outer scope (test runner outside the VM).
   return snapshotState;
