+++ packages/jest-snapshot/package.json
@@ -14,9 +14,13 @@
     "jest-diff": "^23.0.1",
     "jest-matcher-utils": "^23.0.1",
     "jest-message-util": "^23.0.1",
+    "jest-resolve": "^23.0.1",
     "mkdirp": "^0.5.1",
     "natural-compare": "^1.4.0",
     "pretty-format": "^23.0.1",
-    "prettier": "^1.13.4"
+    "semver": "^5.5.0"
+  },
+  "devDependencies": {
+    "prettier": "^1.13.4"     
   }
 }
