+++ packages/jest-snapshot/src/__tests__/utils.test.js
@@ -6,7 +6,6 @@
  */
 
 jest.mock('fs');
-jest.mock('prettier');
 
 const fs = require('fs');
 const path = require('path');
@@ -15,7 +14,6 @@ const chalk = require('chalk');
 const {
   getSnapshotData,
   getSnapshotPath,
-  saveInlineSnapshots,
   keyToTestName,
   saveSnapshotFile,
   serialize,
@@ -28,23 +26,15 @@ const {
 const writeFileSync = fs.writeFileSync;
 const readFileSync = fs.readFileSync;
 const existsSync = fs.existsSync;
-const statSync = fs.statSync;
-const readdirSync = fs.readdirSync;
 beforeEach(() => {
   fs.writeFileSync = jest.fn();
   fs.readFileSync = jest.fn();
   fs.existsSync = jest.fn(() => true);
-  fs.statSync = jest.fn(filePath => ({
-    isDirectory: () => !filePath.endsWith('.js'),
-  }));
-  fs.readdirSync = jest.fn(() => []);
 });
 afterEach(() => {
   fs.writeFileSync = writeFileSync;
   fs.readFileSync = readFileSync;
   fs.existsSync = existsSync;
-  fs.statSync = statSync;
-  fs.readdirSync = readdirSync;
 });
 
 test('keyToTestName()', () => {
@@ -91,100 +81,6 @@ test('saveSnapshotFile() works with \r', () => {
     filename,
     `// Jest Snapshot v1, ${SNAPSHOT_GUIDE_LINK}\n\n` +
       'exports[`myKey`] = `<div>\n</div>`;\n',
-  );
-});
-
-test('saveInlineSnapshots() replaces empty function call with a template literal', () => {
-  const filename = path.join(__dirname, 'my.test.js');
-  fs.readFileSync = jest.fn(() => `expect(1).toMatchInlineSnapshot();\n`);
-
-  saveInlineSnapshots([
-    {
-      frame: {column: 11, file: filename, line: 1},
-      snapshot: `1`,
-    },
-  ]);
-
-  expect(fs.writeFileSync).toHaveBeenCalledWith(
-    filename,
-    'expect(1).toMatchInlineSnapshot(`1`);\n',
-  );
-});
-
-test('saveInlineSnapshots() replaces existing template literal', () => {
-  const filename = path.join(__dirname, 'my.test.js');
-  fs.readFileSync = jest.fn(() => 'expect(1).toMatchInlineSnapshot(`2`);\n');
-
-  saveInlineSnapshots([
-    {
-      frame: {column: 11, file: filename, line: 1},
-      snapshot: `1`,
-    },
-  ]);
-
-  expect(fs.writeFileSync).toHaveBeenCalledWith(
-    filename,
-    'expect(1).toMatchInlineSnapshot(`1`);\n',
-  );
-});
-
-test('saveInlineSnapshots() replaces existing template literal with property matchers', () => {
-  const filename = path.join(__dirname, 'my.test.js');
-  fs.readFileSync = jest.fn(
-    () => 'expect(1).toMatchInlineSnapshot({}, `2`);\n',
-  );
-
-  saveInlineSnapshots([
-    {
-      frame: {column: 11, file: filename, line: 1},
-      snapshot: `1`,
-    },
-  ]);
-
-  expect(fs.writeFileSync).toHaveBeenCalledWith(
-    filename,
-    'expect(1).toMatchInlineSnapshot({}, `1`);\n',
-  );
-});
-
-test('saveInlineSnapshots() throws if frame does not match', () => {
-  const filename = path.join(__dirname, 'my.test.js');
-  fs.readFileSync = jest.fn(() => 'expect(1).toMatchInlineSnapshot();\n');
-
-  const save = () =>
-    saveInlineSnapshots([
-      {
-        frame: {column: 2 /* incorrect */, file: filename, line: 1},
-        snapshot: `1`,
-      },
-    ]);
-
-  expect(save).toThrowError(/Couldn't locate all inline snapshots./);
-});
-
-test('saveInlineSnapshots() throws if multiple calls to to the same location', () => {
-  const filename = path.join(__dirname, 'my.test.js');
-  fs.readFileSync = jest.fn(() => 'expect(1).toMatchInlineSnapshot();\n');
-
-  const frame = {column: 11, file: filename, line: 1};
-  const save = () =>
-    saveInlineSnapshots([{frame, snapshot: `1`}, {frame, snapshot: `2`}]);
-
-  expect(save).toThrowError(
-    /Multiple inline snapshots for the same call are not supported./,
-  );
-});
-
-test('saveInlineSnapshots() uses escaped backticks', () => {
-  const filename = path.join(__dirname, 'my.test.js');
-  fs.readFileSync = jest.fn(() => 'expect("`").toMatchInlineSnapshot();\n');
-
-  const frame = {column: 13, file: filename, line: 1};
-  saveInlineSnapshots([{frame, snapshot: '`'}]);
-
-  expect(fs.writeFileSync).toHaveBeenCalledWith(
-    filename,
-    'expect("`").toMatchInlineSnapshot(`\\``);\n',
   );
 });
