+++ packages/jest-validate/src/__tests__/fixtures/jest_config.js
@@ -41,6 +41,7 @@ const defaultConfig = {
   notify: false,
   notifyMode: 'always',
   preset: null,
+  prettier: 'prettier',
   resetMocks: false,
   resetModules: false,
   restoreMocks: false,
@@ -99,6 +100,7 @@ const validConfig = {
   notify: false,
   notifyMode: 'always',
   preset: 'react-native',
+  prettier: '<rootDir>/node_modules/prettier',
   resetMocks: false,
   resetModules: false,
   restoreMocks: false,
