+++ types/Config.js
@@ -53,6 +53,7 @@ export type DefaultOptions = {|
   notify: boolean,
   notifyMode: string,
   preset: ?string,
+  prettier: string,
   projects: ?Array<string | ProjectConfig>,
   resetMocks: boolean,
   resetModules: boolean,
@@ -135,6 +136,7 @@ export type InitialOptions = {
   passWithNoTests?: boolean,
   preprocessorIgnorePatterns?: Array<Glob>,
   preset?: ?string,
+  prettier?: ?string,
   projects?: Array<Glob>,
   replname?: ?string,
   resetMocks?: boolean,
@@ -258,6 +260,7 @@ export type ProjectConfig = {|
   modulePathIgnorePatterns: Array<string>,
   modulePaths: Array<string>,
   name: string,
+  prettier: string,
   resetMocks: boolean,
   resetModules: boolean,
   resolver: ?Path,
